refactor(canary): Move utility functions to lib directory
- Refactor the project structure to improve organization by moving
  utility functions from various components to a centralized `lib`
  directory. This allows for better separation of concerns and easier
  maintenance of utility functions.

- Update import statements across the project to reflect the new file
  locations.

No breaking changes introduced with this refactor.
