+++ app/(playground)/p/[agentId]/canary/components/editor.tsx
@@ -17,8 +17,8 @@ import { useMousePosition } from "../contexts/mouse-position";
 import { usePropertiesPanel } from "../contexts/properties-panel";
 import { useToast } from "../contexts/toast";
 import { useToolbar } from "../contexts/toolbar";
+import { createNodeId, isTextGeneration } from "../lib/utils";
 import type { NodeId, Tool } from "../types";
-import { createNodeId, isTextGeneration } from "../utils";
 import { Edge } from "./edge";
 import { Header } from "./header";
 import { KeyboardShortcut } from "./keyboard-shortcut";
